refactor catalog
